Make pull request event trigger actions configurable in LabHub TOML file
For some Use Cases GitLab CI pipelines are triggered too often when
changes to a pull request are made in GitHub.
Pull request event trigger actions are now configurable in the LabHub TOML
file.
Only those actions trigger pipelines which are white-listed in the
configuration file.
This solution refers to the issue #107 "pipeline is to often triggered".